requirements.txt
streamlit
bcrypt
pytest
langchain 
langchain-openai 
langchain_core
openai
python-dotenv
python-docx
pdfplumber